requirements.txt
langchain>=0.3.0
openai>=1.50.0
pyyaml>=6.0
pydantic>=2.0.0
loguru>=0.7.0
requests>=2.31.0